+++ CHANGELOG.md
@@ -11,6 +11,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Implement i18n with [lingui](https://lingui.dev).
 
+### Fixed
+
+- Fix a bug in wikidata data parsing.
+
 ## [0.3.0] - 2025-02-19
 
 ### Added